move classes tests
